既に存在している場合はコピーをしないようにする

+++ git/git_template.sh
@@ -1,5 +1,5 @@
 cat <<__EOT__
-$1ディレクトリに以下のテンプレートファイルをコピーします！
+[$1]に以下のテンプレートファイルをコピーします！
 .github/ISSUE_TEMPLATE.md
 .github/PULL_REQUEST_TEMPLATE.md
 .commit_template
@@ -7,7 +7,12 @@ __EOT__
 
 cp -r .github $1
 cp .commit_template $1
-cp .gitignore $1
+if [ -f $1/.gitignore ]; then
+  echo "すでに存在しているので、.gitignoreはコピーしません"
+else
+  echo ".gitignore"
+  cp .gitignore $1
+fi
 # emoji prefix
 cd $1
 git config commit.template .commit_template
